+++ changelog.md
@@ -117,4 +117,5 @@
 
 ## Next
 
+* textSpaced form field format - which preserves line spacing
 * Code tidy